// src/models/payment.model.ts
export interface Payment{

    key: string; 
    user_key: string; 
    chap_key: string; 
    ch_key: string; 
    date: Date;
    buy: boolean;
    by: string;
    title: string; 
    price: number;
    author_key: string;
    id: string;
    paid: number; 
    user_name: string; 
    author_name: string; 
    chap_title: string; 
    chr_title: string;
}